refactor: remove interface to separate folder
