+++ src/app/booking/models/current-flight.model.ts
@@ -1,0 +1,9 @@
+import { IFlightInfo } from 'src/app/shared/models/interfaces';
+
+export interface ICurrentFlightSummary {
+  from: string;
+  to: string;
+  startDate: string;
+  index: number;
+  flightData?: IFlightInfo;
+}
